requirements.txt
Django==2.2.0
django_compressor==2.2
python-gettext==4.0
psycopg2-binary==2.8
pillow==5.4.1
django-modeltranslation==0.13b3
django-suit==0.2.26
dj-database-url==0.5.0
boto3==1.9.98
django-ckeditor==5.6.1
ipdb==0.11.0
django-storages==1.7.1
raven==6.10.0
xhtml2pdf==0.2.3
reportlab==3.6.6
djangorestframework==3.12.1
django-extensions==3.1.0
requests==2.26.0